--- src/main/java/cn/zhouyafeng/itchat4j/components/Contact.java
@@ -1,12 +1,0 @@
-package cn.zhouyafeng.itchat4j.components;
-
-import com.alibaba.fastjson.JSONArray;
-
-import cn.zhouyafeng.itchat4j.utils.Core;
-
-public class Contact {
-	public static Object updateLocalChatrooms(Core core, JSONArray l) {
-		// TODO
-		return null;
-	}
-}
