--- src/main/java/cn/zhouyafeng/itchat4j/utils/Constant.java
@@ -1,6 +1,0 @@
-package cn.zhouyafeng.itchat4j.utils;
-
-public class Constant {
-	public static final String API_WXAPPID = "API_WXAPPID";
-
-}
