+++ src/test/java/cn/zhouyafeng/itchat4j/demo/TulingRobot.java
@@ -15,9 +15,9 @@ import com.alibaba.fastjson.JSONObject;
 import cn.zhouyafeng.itchat4j.Wechat;
 import cn.zhouyafeng.itchat4j.face.IMsgHandlerFace;
 import cn.zhouyafeng.itchat4j.utils.Core;
-import cn.zhouyafeng.itchat4j.utils.DownloadTools;
 import cn.zhouyafeng.itchat4j.utils.MsgType;
 import cn.zhouyafeng.itchat4j.utils.MyHttpClient;
+import cn.zhouyafeng.itchat4j.utils.tools.DownloadTools;
 
 /**
  * 图灵机器人示例
